+++ test/mock-api/reports/remove_report_scope.spec.ts
@@ -1,0 +1,101 @@
+import crypto from 'crypto';
+import { createClient, findWireMockRequest } from '../utils/utils';
+import { expect } from '@jest/globals';
+import {
+    TEST_REPORT_ID,
+    TEST_SHEET_ID,
+    TEST_WORKSPACE_ID,
+    TEST_SUCCESS_MESSAGE,
+    TEST_SUCCESS_RESULT_CODE,
+    ERROR_500_STATUS_CODE,
+    ERROR_500_MESSAGE,
+    ERROR_400_STATUS_CODE,
+    ERROR_400_MESSAGE
+} from './common_test_constants';
+import { ReportAssetType } from '@smartsheet/reports/types';
+
+describe('Reports - removeReportScope endpoint tests', () => {
+    const client = createClient();
+
+    it('removeReportScope generated url is correct', async () => {
+        const requestId = crypto.randomUUID();
+        const options = {
+            reportId: TEST_REPORT_ID,
+            body: [
+                { assetType: ReportAssetType.SHEET, assetId: TEST_SHEET_ID }
+            ],
+            customProperties: {
+                'x-request-id': requestId,
+                'x-test-name': '/reports/remove-report-scope/all-response-body-properties'
+            }
+        };
+        await client.reports.removeReportScope(options);
+        const matchedRequest = await findWireMockRequest(requestId);
+
+        const parsedUrl = new URL(matchedRequest.absoluteUrl);
+        expect(parsedUrl.pathname).toEqual(`/2.0/reports/${TEST_REPORT_ID}/scope`);
+    });
+
+    it('removeReportScope all response body properties', async () => {
+        const requestId = crypto.randomUUID();
+        const options = {
+            reportId: TEST_REPORT_ID,
+            body: [
+                { assetType: ReportAssetType.SHEET, assetId: TEST_SHEET_ID },
+                { assetType: ReportAssetType.WORKSPACE, assetId: TEST_WORKSPACE_ID }
+            ],
+            customProperties: {
+                'x-request-id': requestId,
+                'x-test-name': '/reports/remove-report-scope/all-response-body-properties'
+            }
+        };
+        const response = await client.reports.removeReportScope(options);
+
+        expect(response).toEqual({
+            message: TEST_SUCCESS_MESSAGE,
+            resultCode: TEST_SUCCESS_RESULT_CODE
+        });
+    });
+
+    it('removeReportScope error 500 response', async () => {
+        const requestId = crypto.randomUUID();
+        const options = {
+            reportId: TEST_REPORT_ID,
+            body: [
+                { assetType: ReportAssetType.SHEET, assetId: TEST_SHEET_ID }
+            ],
+            customProperties: {
+                'x-request-id': requestId,
+                'x-test-name': '/errors/500-response'
+            }
+        };
+        try {
+            await client.reports.removeReportScope(options);
+            expect(true).toBe(false); // Expected an error to be thrown
+        } catch (error) {
+            expect(error.statusCode).toBe(ERROR_500_STATUS_CODE);
+            expect(error.message).toBe(ERROR_500_MESSAGE);
+        }
+    });
+
+    it('removeReportScope error 400 response', async () => {
+        const requestId = crypto.randomUUID();
+        const options = {
+            reportId: TEST_REPORT_ID,
+            body: [
+                { assetType: ReportAssetType.SHEET, assetId: TEST_SHEET_ID }
+            ],
+            customProperties: {
+                'x-request-id': requestId,
+                'x-test-name': '/errors/400-response'
+            }
+        };
+        try {
+            await client.reports.removeReportScope(options);
+            expect(true).toBe(false); // Expected an error to be thrown
+        } catch (error) {
+            expect(error.statusCode).toBe(ERROR_400_STATUS_CODE);
+            expect(error.message).toBe(ERROR_400_MESSAGE);
+        }
+    });
+});